Implemented better permissions checks

+++ IssueTracker.Data/Person.cs
@@ -8,7 +8,7 @@
     {
         public long Id { get; set; }
 
-        public int? CustomerId { get; set; }
+        public long? CustomerId { get; set; }
 
         public string Name { get; set; }
